Unit Tests
Create unit tests to ensure units stats are correct.

This is necessary because, as expected, the type data available on the
game assets seens unused and incorrect.

As I'm unable to find the correct data, the only solution is to manually
adjust the correction value until I find a value that works for both old
and new units.